\name{ksmm-package}
\alias{ksmm-package}
\alias{ksmm}
\docType{package}
\title{
 A R Package to solve kernel support matrix machine's problem
}
\description{
 Funtions for Kernel Support Matrix Machine.
}
\details{

}
\author{
Kyuri Park

Maintainer: Kyuri Park <kyuripark12@uos.ac.kr>
}
\references{
 Ye, Y. (2019). A nonlinear kernel support matrix machine. \emph{International Journal of Machine Learning and Cybernetics}. \bold{10}.
}
\keyword{ package }
\examples{
library(devtools)
install_github("kyuridata/ksmm")
library(ksmm)
 
}
